Change data type from double to decimal in Estate properties

+++ Estates/Parcel.cs
@@ -8,7 +8,7 @@ namespace EstateManager.Estates
     {
         public ParcelType ParcelType { get; private set; }
 
-        public Parcel(string adress, double width, double length, double price, OwnerType owner, ParcelType parcelType, DateTime addedDate)
+        public Parcel(string adress, decimal width, decimal length, decimal price, OwnerType owner, ParcelType parcelType, DateTime addedDate)
             : base(adress, width, length, price, owner, addedDate)
         {
             ParcelType = parcelType;
